python/requirements.txt
youtube_transcript_api
google-generativeai
python-dotenv
pathlib
Flask